reorganised to follow golang-standards/project-layout
