+++ go.mod
@@ -1,7 +1,6 @@
-module portablelan
+module yapl
 
 go 1.24.6
 
 require github.com/ulikunitz/xz v0.5.15
-
-require github.com/klauspost/compress v1.18.0 // indirect
+require github.com/klauspost/compress v1.18.0
